+++ src/scriptlets/trusted-set-cookie-reload.js
@@ -3,7 +3,7 @@ import {
     logMessage,
     nativeIsNaN,
     isCookieSetWithValue,
-    concatCookieNameValuePath,
+    serializeCookie,
     isValidCookiePath,
     parseKeywordValue,
     getTrustedCookieOffsetMs,
@@ -19,14 +19,15 @@ import {
  *
  * @description
  * Sets a cookie with arbitrary name and value,
- * and with optional ability to offset cookie attribute 'expires' and set path.
+ * and with optional ability to offset cookie attribute 'expires', set path
+ * and set domain.
  * Also reloads the current page after the cookie setting.
  * If reloading option is not needed, use the [`trusted-set-cookie` scriptlet](#trusted-set-cookie).
  *
  * ### Syntax
  *
  * ```text
- * example.org#%#//scriptlet('trusted-set-cookie-reload', name, value[, offsetExpiresSec[, path]])
+ * example.org#%#//scriptlet('trusted-set-cookie-reload', name, value[, offsetExpiresSec[, path[, domain]]])
  * ```
  *
  * - `name` — required, cookie name to be set
@@ -43,6 +44,8 @@ import {
  * - `path` — optional, argument for setting cookie path, defaults to `/`; possible values:
  *     - `/` — root path
  *     - `none` — to set no path at all
+ * - `domain` — optional, cookie domain, if not set origin will be set as domain,
+ *              if the domain does not match the origin, the cookie will not be set
  *
  * > Note that the scriptlet does not encode cookie names and values.
  * > As a result, if a cookie's name or value includes `;`,
@@ -80,11 +83,17 @@ import {
  *     example.org#%#//scriptlet('trusted-set-cookie-reload', 'cmpconsent', 'decline', '', 'none')
  *     ```
  *
+ * 1. Set cookie with domain
+ *
+ *     ```adblock
+ *     example.org#%#//scriptlet('trusted-set-cookie-reload', 'cmpconsent', 'decline', '', 'none', 'example.org')
+ *     ```
+ *
  * @added v1.7.10.
  */
 /* eslint-enable max-len */
 
-export function trustedSetCookieReload(source, name, value, offsetExpiresSec = '', path = '/') {
+export function trustedSetCookieReload(source, name, value, offsetExpiresSec = '', path = '/', domain = '') {
     if (typeof name === 'undefined') {
         logMessage(source, 'Cookie name should be specified');
         return;
@@ -107,12 +116,18 @@ export function trustedSetCookieReload(source, name, value, offsetExpiresSec = '
         return;
     }
 
-    let cookieToSet = concatCookieNameValuePath(name, parsedValue, path, false);
+    if (!document.location.origin.includes(domain)) {
+        logMessage(source, `Cookie domain not matched by origin: '${domain}'`);
+        return;
+    }
+
+    let cookieToSet = serializeCookie(name, parsedValue, path, domain, false);
     if (!cookieToSet) {
         logMessage(source, 'Invalid cookie name or value');
         return;
     }
 
+    // TODO: Move this concat to serializeCookie
     if (offsetExpiresSec) {
         const parsedOffsetMs = getTrustedCookieOffsetMs(offsetExpiresSec);
 
@@ -150,7 +165,7 @@ trustedSetCookieReload.injections = [
     logMessage,
     nativeIsNaN,
     isCookieSetWithValue,
-    concatCookieNameValuePath,
+    serializeCookie,
     isValidCookiePath,
     getTrustedCookieOffsetMs,
     parseKeywordValue,