+++ tests/helpers/cookie-utils.spec.js
@@ -1,6 +1,6 @@
-import { concatCookieNameValuePath } from '../../src/helpers/cookie-utils';
+import { serializeCookie } from '../../src/helpers/cookie-utils';
 
-describe('concatCookieNameValuePath', () => {
+describe('serializeCookie', () => {
     describe('encode cookie value', () => {
         test.each([
             {
@@ -37,44 +37,49 @@ describe('concatCookieNameValuePath', () => {
                 actual: ['abc', 'de;f', ''],
                 expected: 'abc=de%3Bf',
             },
+            // set domain
+            {
+                actual: ['test', '1', '', 'example.com'],
+                expected: 'test=1; domain=example.com',
+            },
         ])('$actual -> $expected', ({ actual, expected }) => {
-            expect(concatCookieNameValuePath(...actual)).toBe(expected);
+            expect(serializeCookie(...actual)).toBe(expected);
         });
     });
 
     describe('no cookie value encoding', () => {
         test.each([
             {
-                actual: ['name', 'value', ''],
+                actual: ['name', 'value', '', '', false],
                 expected: 'name=value',
             },
             {
-                actual: ['__test-cookie_expires', 'expires', '/'],
+                actual: ['__test-cookie_expires', 'expires', '/', '', false],
                 expected: '__test-cookie_expires=expires; path=/',
             },
             {
-                actual: ['aa::bb::cc', '1', ''],
+                actual: ['aa::bb::cc', '1', '', '', false],
                 expected: 'aa::bb::cc=1',
             },
             {
-                actual: ['__w_cc11', '{%22cookies_statistical%22:false%2C%22cookies_ad%22:true}', ''],
+                actual: ['__w_cc11', '{%22cookies_statistical%22:false%2C%22cookies_ad%22:true}', '', '', false],
                 // do not encode cookie value
                 // https://github.com/AdguardTeam/Scriptlets/issues/311
                 expected: '__w_cc11={%22cookies_statistical%22:false%2C%22cookies_ad%22:true}',
             },
             // invalid name because of ';'
             {
-                actual: ['a;bc', 'def', ''],
+                actual: ['a;bc', 'def', '', '', false],
                 expected: null,
             },
             // invalid value because of ';' and it is not being encoded
             {
-                actual: ['abc', 'de;f', ''],
+                actual: ['abc', 'de;f', '', '', false],
                 expected: null,
             },
         ])('$actual -> $expected', ({ actual, expected }) => {
             // explicit 'false' to disable encoding
-            expect(concatCookieNameValuePath(...actual, false)).toBe(expected);
+            expect(serializeCookie(...actual)).toBe(expected);
         });
     });
 });